feat: 라이선스 가져오기 API 정렬 조건 추가

+++ license-service/src/main/java/com/teamdev/licenseservice/dto/LicenseWithProductCountDto.java
@@ -22,10 +22,10 @@ public class LicenseWithProductCountDto {
 
     @NotNull
     @PositiveOrZero
-    private Integer expiredProductCount;
+    private Long expiredProductCount;
 
     @Builder
-    public LicenseWithProductCountDto(String key, Long totalProductCount, Integer expiredProductCount) {
+    public LicenseWithProductCountDto(String key, Long totalProductCount, Long expiredProductCount) {
         this.key = key;
         this.totalProductCount = totalProductCount;
         this.expiredProductCount = expiredProductCount;
